Copy code slice from nci to oci
This appears to be a basic error in converting Lua 5.2 from C to Go.
Since C allows pointer arithmetic, it made sense to just have savedpc
be a pointer into the instruction table and increment it as needed. So,
when Lua in C copies savedpc from nci to oci, that's equivalent to
taking copying both nci.savedPC and nci.code in Go.

Closes #46.