Add tests for userdata/table nil comparison
TestTableEqualityNil borrowed from #51 by @gdmen with an additional
assert to check that the comparison returns a boolean false (no reason
to believe it wouldn't, but might as well check it).

TestUserDataEqualityNil has to go a step farther and push userdata, so
it's a bit longer.

TestTableUserdataEquality tries to test for a panic on comparing
the different types with metatables.