Simplify type check on (*State).equalObjects
Removes excess `return false`. This also removes the ok check since we
only really care if t2 is a table or userdata. If t2 is also a nil
pointer to a table/UD, then the Lua state is somehow far worse off than
anticipated anyway and a panic may be more valuable when accessing
t2.metaTable.