Fix plain argument check for string.find
plain defaults to true, but (*State).ToBoolean returns false if the
value is out of range. So, for cases where plain isn't provided (i.e.,
the type of the value is none), plain is true. Otherwise, if the
argument is provided, use the result of (*State).ToBoolean.

Fixes secondary issue introduced by #68.