Get rid of default closure in match function
The original version of match() used goto labels to avoid recursive
function calls and keep control over the stack size.  However, Go cannot
jump from one block to another, so the original location of the "dflt"
label was unusable.  Until this commit, I worked around this deficiency
with a closure, but this made the code harder to follow and
"out-of-order" compared to the original codebase.

The closure is gone, and instead we move the default case of the main
switch out into the parent block, reintroducing the old "dflt" label.
In cases where we do not want to execute the "dflt" label, we skip over
it with a new label that goes straight to the "end" of the function.