Avoid undefined behaviour when doing range checks for integers
Casting from floating point to integer is undefined for values that
exceed the valid range of the target integer type and this results in
failure of test `assert(not pcall(string.format, "%d", 2^63))` in
lua-tests for aarch64.

Replace with a safer test for value ranges that does not rely on the
cast from float to integer.